Les classes chargées d'afficher les vues ont été déplacés dans le package vue.screen
