+++ core/src/lrp/mygdx/game/MyGdxGame.java
@@ -1,7 +1,8 @@
 package lrp.mygdx.game;
 
 import com.badlogic.gdx.Game;
-import vue.sceens.ImageLancement;
+
+import vue.screen.ImageLancement;
