+++ core/src/vue/screen/ImageLancement.java
@@ -1,4 +1,4 @@
-package vue.sceens;
+package vue.screen;
 
 
 import lrp.mygdx.game.MyGdxGame;
