+++ core/src/vue/screen/Menu.java
@@ -1,4 +1,4 @@
-package vue.sceens;
+package vue.screen;
